feat(adapter): modify the pulumi adapter to enable it to work within the state directory
