+++ components/adapters/pulumi/src/utils/tools.ts
@@ -1,0 +1,3 @@
+export function sleep(sec: number) {
+  return new Promise((resolve) => setTimeout(resolve, sec * 1000));
+}
